requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-richtextfield==1.6.1
idna==3.4
psycopg2==2.9.5
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.13